requirements.txt
apache-airflow==2.5.3
apache-airflow-providers-ssh
apache-airflow-providers-hashicorp
boto3==1.24.32
plyvel
webdavclient3
--index-url https://gitlab.jsc.fz-juelich.de/api/v4/projects/4405/packages/pypi/simple
airflow-datacat-integration>=0.1.4
virtualenv
pyunicore==0.15.0
python-gitlab==3.15.0